Add a rudimentary entity system
