+++ mips/game.py
@@ -1,12 +1,7 @@
-from random import random
-
 import pygame, sys, time
-from pygame.locals import QUIT, KEYDOWN, K_LEFT, K_RIGHT, K_UP, K_DOWN
+from pygame.locals import QUIT
 
-
-def mistake():
-    """Return a random number between -5 and 5."""
-    return random()*10 - 5
+from mips.vehicle import Vehicle
 
 
 FPS = 30
@@ -30,33 +25,26 @@ def run():
     pygame.display.set_caption('Animation')
     background = pygame.image.load('bg.png')
 
-    sprite = pygame.image.load('down.png')
-    spritex = 0
-    spritey = 0
+    drawables = []
+    updatables = []
+
+    vehicle = Vehicle('down.png', x=0, y=0, surface=displaysurf)
+    drawables.append(vehicle)
+    updatables.append(vehicle)
 
     while True:
         displaysurf.blit(background, (0, 0))
 
-        displaysurf.blit(sprite, (spritex, spritey))
+        for drawable in drawables:
+            drawable.draw()
 
         for event in pygame.event.get():
+            for updatable in updatables:
+                updatable.update(event)
+
             if event.type == QUIT:
                 pygame.quit()
                 sys.exit()
-
-            if event.type == KEYDOWN:
-                if event.key == K_LEFT:
-                    spritex -= 70 + mistake()
-                    # sprite=pygame.image.load('left.png')
-                elif event.key == K_RIGHT:
-                    spritex += 70 + mistake()
-                    # sprite=pygame.image.load('right.png')
-                elif event.key == K_UP:
-                    spritey -= 50 + mistake()
-                    # sprite=pygame.image.load('up.png')
-                elif event.key == K_DOWN:
-                    spritey += 50 + mistake()
-                    # sprite=pygame.image.load('down.png')
 
         pygame.display.update()
         fps_clock.tick(FPS)
