+++ mips/utils.py
@@ -1,0 +1,6 @@
+from random import random
+
+
+def mistake():
+    """Return a random number between -5 and 5."""
+    return random()*10 - 5
